+++ src/main/java/com/github/ktoublanc/visual/regressions/diff/AverageImageDifferences.java
@@ -1,9 +1,9 @@
 package com.github.ktoublanc.visual.regressions.diff;
 
+import com.github.ktoublanc.visual.regressions.model.Exclusion;
+
 import java.awt.image.BufferedImage;
 import java.util.List;
-
-import com.github.ktoublanc.visual.regressions.model.Exclusion;
 
 /**
  * Average image difference implementation.
@@ -17,15 +17,14 @@ public class AverageImageDifferences extends AbstractImageDifferences implements
 
 	/**
 	 * Builds a {@link ImageDifferences} class instances specifying the compared images and the analysing rectangle size
-	 *
-	 * @param referenceImage The reference image
+	 *  @param referenceImage The reference image
 	 * @param comparedImage  The compared image
 	 * @param threshold      The difference threshold
 	 * @param size           The analysing rectangle size
 	 * @param exclusions     The excluded part of the image
 	 */
 	public AverageImageDifferences(final BufferedImage referenceImage, final BufferedImage comparedImage,
-	                               final int threshold, final int size, final List<Exclusion> exclusions) {
+	                               final double threshold, final int size, final List<Exclusion> exclusions) {
 		super(referenceImage, comparedImage, threshold, exclusions);
 		this.size = size < 0 ? DEFAULT_RECT_SIZE : size;
 	}
